refactor: remove unused timeline component and move the timeline component
