+++ src/components/JobsHistory/Filters/JobsHistoryFilters.tsx
@@ -1,6 +1,6 @@
 import FormikFilterForm from "@flanksource-ui/components/Forms/FormikFilterForm";
-import { TimeRangePicker } from "@flanksource-ui/ui/TimeRangePicker";
-import useTimeRangeParams from "@flanksource-ui/ui/TimeRangePicker/useTimeRangeParams";
+import { TimeRangePicker } from "@flanksource-ui/ui/Dates/TimeRangePicker";
+import useTimeRangeParams from "@flanksource-ui/ui/Dates/TimeRangePicker/useTimeRangeParams";
 import { URLSearchParamsInit } from "react-router-dom";
 import JobHistoryDurationDropdown from "./JobHistoryDurationDropdown";
 import JobHistoryNamesDropdown from "./JobHistoryNames";
