+++ src/ui/Dates/TimeRangePicker/TimeRangePicker.stories.tsx
@@ -1,12 +1,22 @@
 import { ComponentMeta, ComponentStory } from "@storybook/react";
+import { useState } from "react";
 import { TimeRangePicker } from "./index";
 import { TimeRangeOption } from "./rangeOptions";
 
 const TimeRangePickerContainer = () => {
+  const [value, setValue] = useState<TimeRangeOption | undefined>();
+
   const onChange = (range: TimeRangeOption) => {
     console.log("range", range);
+    setValue(range);
   };
-  return <TimeRangePicker style={{ width: "450px" }} onChange={onChange} />;
+  return (
+    <TimeRangePicker
+      style={{ width: "450px" }}
+      onChange={onChange}
+      value={value}
+    />
+  );
 };
 
 export default {
